+++ src/pages/Profile/index.ts
